Updated frontend(re-structured the files),removed few redundant old files
