+++ frontend/src/components/coding/TestRunner.jsx
@@ -1,0 +1,9 @@
+// src/components/coding/TestRunner.jsx
+export default function TestRunner() {
+  return (
+    <div style={{ padding: 20, textAlign: 'center', color: '#6c757d' }}>
+      <h3>Test Runner</h3>
+      <p>Coming soon: Automated test case execution for coding challenges</p>
+    </div>
+  );
+}
