--- frontend/src/pages/ResumeAnalyzerPage.jsx
@@ -1,12 +1,0 @@
-// frontend/src/pages/ResumeAnalyzerPage.jsx
-
-import React from 'react';
-import ResumeAnalyzer from '../components/candidate/ResumeAnalyzer/ResumeAnalyzer';
-
-export default function ResumeAnalyzerPage() {
-  return (
-    <div style={{ minHeight: '100vh', backgroundColor: '#f5f5f5' }}>
-      <ResumeAnalyzer />
-    </div>
-  );
-}
